+++ packages/investor-yearn/src/api/vaults.ts
@@ -1,7 +1,6 @@
 import { ChainTypes } from '@shapeshiftoss/types'
-import { Vault } from '@yfi/sdk'
 
-import { transformVault } from '../utils/transformVault'
+import { transformVault } from '../utils'
 import { YearnVault } from './api'
 import { yearnSdk } from './yearn-sdk'
 
@@ -18,5 +17,5 @@ export type SupportedYearnVault = YearnVault & {
 
 export const getSupportedVaults = async (): Promise<SupportedYearnVault[]> => {
   const vaults = await yearnSdk.vaults.get()
-  return vaults.map((vault: Vault) => transformVault(vault))
+  return vaults.map(transformVault)
 }